refactor: interfaces decoupling
