+++ src/interfaces/journal.interface.ts
@@ -1,0 +1,7 @@
+import { Metric } from "@/interfaces/metric.interface";
+
+export interface Journal{
+  id:number;
+  name:string;
+  metrics:Metric[];
+}
