+++ src/interfaces/metric.interface.ts
@@ -1,0 +1,7 @@
+export interface Metric{
+  id:number;
+  name:string;
+  score:string;
+  year:number;
+  source:string;
+}
